Move `ErasingOp` into `Operators` lint pass
