+++ clippy_lints/src/lib.register_lints.rs
@@ -141,7 +141,6 @@ store.register_lints(&[
     enum_variants::MODULE_INCEPTION,
     enum_variants::MODULE_NAME_REPETITIONS,
     equatable_if_let::EQUATABLE_IF_LET,
-    erasing_op::ERASING_OP,
     escape::BOXED_LOCAL,
     eta_reduction::REDUNDANT_CLOSURE,
     eta_reduction::REDUNDANT_CLOSURE_FOR_METHOD_CALLS,
@@ -431,6 +430,7 @@ store.register_lints(&[
     operators::DOUBLE_COMPARISONS,
     operators::DURATION_SUBSEC,
     operators::EQ_OP,
+    operators::ERASING_OP,
     operators::FLOAT_ARITHMETIC,
     operators::INEFFECTIVE_BIT_MASK,
     operators::INTEGER_ARITHMETIC,
